Updated empty example project_{device}.js files
The user is told to copy project_{device}.js files when creating a new
PRU project, however "empty" is hardcoded in post build steps -- leading
to conflicts when attempting to build and run. I changed this to make
define project_name globally and then reference this in post build steps
and in property.name then instructed user is instructed to change project_name in the top level readme
when making their new project.

Signed-off-by: baileylooper-ti <b-looper@ti.com>