examples/fft: remove Windows make dependency
When building in Windows, the assembly file can do a path with forward slashes.
like this:
.include "path\to\included_file.h"

However, when building in Linux, the assembly file needs to use back slashes.
Like this:
.include "path/to/included_file.h"

However, if the path is provided in the makefile, then the same path can be used
in both Windows and Linux.

This include path is actually already defined in the makefile, so we can remove
the path from the assembly code and just include the file.
Like this:
.include "included_file.h"

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>