academy/crc/crc: add project Makefile
This makefile allows us to build all project files with a single
make command. The top-level makefile can also be used to build all
projects that use a specific processor by calling the project makefile
for every single project.

If the processor in imports.mak > DEVICE is in the SUPPORTED_PROCESSOR list,
then the project makefile will run.

If the processor in imports.mak > DEVICE is not in the SUPPORTED_PROCESSOR list,
then the project makefile silently does nothing.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>