empty/firmware/makefile: use CGT_TI_PRU_PATH
CGT_TI_PRU_PATH is renamed to CG_TOOL_ROOT in the makefile.

This might make sense for autogen tools, but it makes the makefile harder
for a human to understand. Get rid of CG_TOOL_ROOT and use the variable that
a customer is supposed to set in imports.mak.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>