example.projectspec: remove MCU_PLUS_SDK_AMXXX
By default, we want the open-pru repo's projects to be compatible with the
processor-specific version of MCU+ SDK that is packaged and distributed through
the ti.com product page.

MCU_PLUS_SDK_AMXXX is used by the MCU+ SDK git repo. Remove this as the default
option. If customers want to modify their example.projectspec files to enable
development on the MCU+ SDK git repo, we can document those steps separately.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>