uart_12mb v1.2: increase TX buffers from 2->8
Increase FW revision to 0x00010002

In order to enable more than 2 TX buffers, keep track of the last buffer that was
transmitted (PREV_BUF). In IDLE, check the next buffer in the sequence
to see if additional data is ready to send (NEXT_BUF). The PRU's
partner who is providing data should always fill the buffers in sequence in
order to avoid garbling the UART output messages, so we do not need to check
any of the other buffers.

Signed-off-by: Nick Saulnier <nsaulnier@ti.com>